+++ cigionline/static/js/components/EventListingCard.js
@@ -142,7 +142,7 @@ const EventListingCard = (props) => {
                     <i className="fab fa-facebook-f"></i>
                     Share on Facebook
                   </a>
-                  {row.event_access !== 'Private' && (
+                  {row.event_access !== 'Private' && row.registration_url && (
                     <a className="dropdown-item" href={row.registration_url} onClick="ga('send', 'event', 'Event Registration', 'Click' );">
                       <i className="fal fa-check-square"></i>
                       Register
