+++ package.json
@@ -27,7 +27,8 @@
     "d3": "^7.8.0",
     "jquery": "3.6.0",
     "normalize.css": "8.0.1",
-    "react-calendar": "3.4.0",
+    "react-calendar": "4.2.1",
+    "react-ga": "^3.3.1",
     "react-router-dom": "^5.2.0",
     "react-transition-group": "^4.4.2",
     "rimraf": "3.0.2",
